move head count to canteen service
